feat: 돌을 착수하고 오목 및 승리 여부를 확인하는 Player
- 플레이어는 색상을 가지고 있다.
- 플레이어는 돌을 착수할 수 있다.
- 플레이어는 오목 여부를 확인할 수 있다
- 오목, 장목이 되면 해당 플레이어는 승리 상태로 바꾼다
