+++ README.md
@@ -9,9 +9,9 @@
       - [x] 이미 착수된 위치에 돌을 놓을 수 없다.
       - [x] 놓고자 하는 위치가 이미 돌이 착수된 위치라면, `IllegalStateException`를 던져야 한다.
 - 두명의 플레이어가 있다.
-    - [ ] 흑돌을 가진 플레이어와 백돌을 가진 플레이어가 있다.
+    - [x] 흑돌을 가진 플레이어와 백돌을 가진 플레이어가 있다.
     - [ ] 흑돌을 가진 플레이어가 먼저 착수한다.
-    - [ ] 돌을 바둑판 위에 원하는 위치로 착수한다.
+    - [x] 돌을 바둑판 위에 원하는 위치로 착수한다.
 - 돌이 있다.
     - [x] 흑돌과 백돌로 구분된다.
     - [x] 돌은 바둑판 위의 위치를 가지고 있다.
