+++ src/test/kotlin/omok/model/PlayerTest.kt
@@ -4,26 +4,6 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
-class Player(val color: Color) {
-    private var _isWin = false
-    val isWin: Boolean
-        get() = _isWin
-
-    fun playTurn(board: Board, coordinate: Coordinate) {
-        val stone = Stone(color, coordinate)
-        putStoneOnBoard(board, stone)
-        _isWin = checkOmok(board.stones, stone)
-    }
-
-    private fun putStoneOnBoard(board: Board, stone: Stone) {
-        board.putStone(stone)
-    }
-
-    private fun checkOmok(stones: Stones, stone: Stone): Boolean {
-        return stones.findOmok(stone)
-    }
-}
-
 class PlayerTest {
     private lateinit var board: Board
     private lateinit var stones: Stones
@@ -37,7 +17,6 @@ class PlayerTest {
         stones.putStone(Stone(Color.BLACK, Coordinate(Row.from("7"), Column.from("G"))))
         stones.putStone(Stone(Color.BLACK, Coordinate(Row.from("8"), Column.from("H"))))
         stones.putStone(Stone(Color.BLACK, Coordinate(Row.from("9"), Column.from("I"))))
-
     }
 
     @Test
